# server5/src/openapi_server/models/appointment_for_update_dto.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class AppointmentForUpdateDto(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    AppointmentForUpdateDto - a model defined in OpenAPI

        id: The id of this AppointmentForUpdateDto.
        appointment_type: The appointment_type of this AppointmentForUpdateDto [Optional].
        appointment_status: The appointment_status of this AppointmentForUpdateDto [Optional].
        appointment_date: The appointment_date of this AppointmentForUpdateDto [Optional].
        slot_time: The slot_time of this AppointmentForUpdateDto [Optional].
        created_at: The created_at of this AppointmentForUpdateDto [Optional].
        created_by: The created_by of this AppointmentForUpdateDto [Optional].
        doctor_id: The doctor_id of this AppointmentForUpdateDto [Optional].
        symptoms: The symptoms of this AppointmentForUpdateDto [Optional].
        updated_at: The updated_at of this AppointmentForUpdateDto [Optional].
        slot_id: The slot_id of this AppointmentForUpdateDto [Optional].
        patient_id: The patient_id of this AppointmentForUpdateDto [Optional].
        doctor_name: The doctor_name of this AppointmentForUpdateDto [Optional].
        doctor_image: The doctor_image of this AppointmentForUpdateDto [Optional].
        is_exist: The is_exist of this AppointmentForUpdateDto [Optional].
        patient_image: The patient_image of this AppointmentForUpdateDto [Optional].
        patient_name: The patient_name of this AppointmentForUpdateDto [Optional].
        connecty_cube_id: The connecty_cube_id of this AppointmentForUpdateDto [Optional].
    """

    id: str = Field(alias="id")
    appointment_type: Optional[int] = Field(alias="appointmentType", default=None)
    appointment_status: Optional[bool] = Field(alias="appointmentStatus", default=None)
    appointment_date: Optional[date] = Field(alias="appointmentDate", default=None)
    slot_time: Optional[str] = Field(alias="slotTime", default=None)
    created_at: Optional[datetime] = Field(alias="createdAt", default=None)
    created_by: Optional[str] = Field(alias="createdBy", default=None)
    doctor_id: Optional[str] = Field(alias="doctorId", default=None)
    symptoms: Optional[str] = Field(alias="symptoms", default=None)
    updated_at: Optional[datetime] = Field(alias="updatedAt", default=None)
    slot_id: Optional[str] = Field(alias="slotId", default=None)
    patient_id: Optional[str] = Field(alias="patientId", default=None)
    doctor_name: Optional[str] = Field(alias="doctorName", default=None)
    doctor_image: Optional[str] = Field(alias="doctorImage", default=None)
    is_exist: Optional[bool] = Field(alias="isExist", default=None)
    patient_image: Optional[str] = Field(alias="patientImage", default=None)
    patient_name: Optional[str] = Field(alias="patientName", default=None)
    connecty_cube_id: Optional[str] = Field(alias="connectyCubeId", default=None)

AppointmentForUpdateDto.update_forward_refs()