# server5/src/openapi_server/models/appointment_paged_result_dto.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.appointment_dto import AppointmentDto


class AppointmentPagedResultDto(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    AppointmentPagedResultDto - a model defined in OpenAPI

        total_count: The total_count of this AppointmentPagedResultDto [Optional].
        items: The items of this AppointmentPagedResultDto [Optional].
    """

    total_count: Optional[int] = Field(alias="totalCount", default=None)
    items: Optional[List[AppointmentDto]] = Field(alias="items", default=None)

AppointmentPagedResultDto.update_forward_refs()